--- src/main/java/kr/co/glnt/relay/breaker/exception/GlntBadRequestException.java
@@ -1,8 +1,0 @@
-package kr.co.glnt.relay.breaker.exception;
-
-public class GlntBadRequestException extends RuntimeException {
-    public GlntBadRequestException(String msg) {
-        super(msg);
-    }
-
-}
