--- src/main/java/kr/co/glnt/relay/breaker/service/BreakerFactory.java
@@ -1,23 +1,0 @@
-package kr.co.glnt.relay.breaker.service;
-
-import kr.co.glnt.relay.breaker.dto.FacilityInfo;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-public enum BreakerFactory {
-    OUTFRONT, INBACK, INFRONT;
-
-    public static Breaker getInstance(FacilityInfo facilityInfo) {
-        String type = facilityInfo.getGateLprType();
-        switch (BreakerFactory.valueOf(type)) {
-            case INFRONT:
-                return new Entrance(facilityInfo);
-            case INBACK:
-                return new EntranceBack(facilityInfo);
-            case OUTFRONT:
-                return new Exit(facilityInfo);
-            default:
-                return null;
-        }
-    }
-}
