+++ src/main/java/kr/co/glnt/relay/exception/GlntBadRequestException.java
@@ -1,0 +1,8 @@
+package kr.co.glnt.relay.exception;
+
+public class GlntBadRequestException extends RuntimeException {
+    public GlntBadRequestException(String msg) {
+        super(msg);
+    }
+
+}
