--- src/main/java/kr/co/glnt/relay/run/HealthCheckRunner.java
@@ -1,4 +1,0 @@
-package kr.co.glnt.relay.run;
-
-public class HealthCheckRunner {
-}
