+++ src/main/java/kr/co/glnt/relay/service/Breaker.java
@@ -1,0 +1,112 @@
+package kr.co.glnt.relay.service;
+
+
+import kr.co.glnt.relay.dto.CarInfo;
+import kr.co.glnt.relay.dto.EventInfoGroup;
+import kr.co.glnt.relay.web.GpmsAPI;
+import kr.co.glnt.relay.web.NgisAPI;
+import kr.co.glnt.relay.config.ApplicationContextProvider;
+import kr.co.glnt.relay.dto.EventInfo;
+import kr.co.glnt.relay.dto.FacilityInfo;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.ApplicationContext;
+
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+// 차단기 부모.
+@Slf4j
+public abstract class Breaker {
+    protected long lastRegTime = System.currentTimeMillis();
+    protected final int MAX_TIME = 1000;    // 발생하는 이벤트들을 하나로 취급할 시간
+    protected final int TIMER_TIME = 1000;
+    // 주차장 정보. (차단기, 정산기 등..)
+    protected FacilityInfo facilityInfo;
+    protected NgisAPI ngisAPI;
+    protected GpmsAPI gpmsAPI;
+
+    public Breaker(FacilityInfo facilityInfo) {
+        this.facilityInfo = facilityInfo;
+
+        ApplicationContext ctx = ApplicationContextProvider.getApplicationContext();
+        ngisAPI = ctx.getBean("ngisAPI", NgisAPI.class);
+        gpmsAPI = ctx.getBean("gpmsAPI", GpmsAPI.class);
+    }
+
+    /**
+     * 차단기 시작
+     */
+    public abstract void startProcessing(EventInfo eventInfo);
+
+
+    /**
+     * 지정한 시간내에 들어온 차량인지 확인 (같은 차량)
+     */
+    protected boolean isNewCarEnters(long currentTime) {
+        return (currentTime - lastRegTime) > MAX_TIME;
+    }
+
+
+    /**
+     * 이벤트 그룹의 차량 정보들을 OCR 처리하여 차량정보 리스트를 반환
+     */
+    protected List<CarInfo> getCurrentGroupEventList(EventInfoGroup eventGroup) {
+        return eventGroup.getEventList().stream()
+                .map(eventInfo -> {
+                    CarInfo carInfo = ngisAPI.requestOCR(eventInfo.getFullPath());
+                    carInfo.setInDate(eventInfo.getCreatedTime());
+                    carInfo.setFullPath(eventInfo.getFullPath());
+                    carInfo.setFacilitiesId(facilityInfo.getFacilitiesId());
+                    return carInfo;
+                })
+                .collect(Collectors.toList());
+    }
+
+    /**
+     * 입차 요청할 차량정보 추출
+     */
+    protected CarInfo getCarInfoToBeTransmit(List<CarInfo> carInfos) {
+        long ocrCount = carInfos.stream().filter(CarInfo::isState).count();
+        if (ocrCount > 0) {
+            return recognizedVehicle(carInfos);
+        }
+        else {
+            return partiallyRecognizedVehicle(carInfos);
+        }
+    }
+
+    /**
+     *  정상인식 된 차량들 중에 더 많이 찍힌 차량번호를 리턴.
+     */
+    private CarInfo recognizedVehicle(List<CarInfo> carInfos) {
+        String carNumber = null;
+        // OCR 이 인식된 값들을 차량번호를 키로 하여 map으로 변환.
+        Map<String, List<CarInfo>> carMap = carInfos.stream()
+                .filter(CarInfo::isState)
+                .collect(Collectors.groupingBy(CarInfo::getNumber));
+
+        // 인식된 번호중 더 많이 찍힌 차량번호를 추출.
+        carNumber = carMap.entrySet().stream()
+                .max((entry1, entry2) -> entry1.getValue().size() > entry2.getValue().size() ? 1 : -1)
+                .get()
+                .getKey();
+
+        // 차량 정보 가져오기
+        return carMap.get(carNumber).get(0);
+    }
+
+    /**
+     * 미인식된 차량들 중에 부분인식 정보가 있으면 해당 정보를 리턴하고
+     * 없을 경우 첫번째 미인식 정보를 리턴한다.
+     */
+    private CarInfo partiallyRecognizedVehicle(List<CarInfo> carInfos) {
+        for (CarInfo info : carInfos) {
+            if (info.getCode() > 10) { // 부분인식이 있으면
+                return info;
+            }
+        }
+        return carInfos.get(0);
+    }
+
+}
