+++ src/main/resources/application.yml
@@ -15,3 +15,4 @@ spring:
 
   mvc:
     throw-exception-if-no-handler-found: true
+
